% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{clusterAnalysisTwo}
\alias{clusterAnalysisTwo}
\title{Applies Clustering to DNA Methylation Data using the mclust package}
\usage{
clusterAnalysisTwo(dnaMethylationData, dnaMethylationSites)
}
\arguments{
\item{dnaMethylationData}{A matrix which contains DNA Methylation beta
values of various methylation sites from several different samples.}

\item{dnaMethylationSites}{A matrix that contains the DNA methylation sites
which have been validated as being significant for the purposes of
clustering. These are to be generated using cox regression analysis.}
}
\value{
Returns a list with the results.
\itemize{
\item call - The matched call.
\item data - Input data matrix.
\item modelName - Character string denoting the model.
\item n - Number of observations in data.
\item d - dimension of the data.
\item G - Optimal number of mixture components.
\item BIC - All BIC values.
\item loglik -  log-likelihood corresponding to the optimal BIC.
\item df - The number of estimated parameters.
\item bic - BIC value of the selected model.
\item icl - ICL value of the selected model.
\item hypvol - The hypervolume parameter.
\item parameters - A list with the following: pro, mean and variance.
\item z - A matrix with probabilities.
\item classification - The classification corresponding to z.
\item uncertainty - The uncertainty.
}
}
\description{
A function that does clustering, using mclust package, on DNA methylation
Data taken from  multiple samples given the significant methylation sites.
}
\examples{
# Example 1:
# Using methylationData and significantSites Data available with package
dim(methylationData)
dim(significantSites)

# Do clustering Analysis
results <- clusterAnalysisTwo(dnaMethylationData = methylationData, dnaMethylationSites = significantSites)
results$classification

}
\references{
Wickham H, François R, Henry L, Müller K (2022). \emph{dplyr: A Grammar of Data
Manipulation}. R package version 1.0.10, \url{https://CRAN.R-project.org/package=dplyr}.

R Core Team (2022). R: A language and environment for statistical computing. R
Foundation for Statistical Computing, Vienna, Austria. URL
https://www.R-project.org/.

Li, T., Chen, X., Gu, M. et al. Identification of the subtypes of gastric
cancer based on DNA methylation and the prediction of prognosis.
Clin Epigenet 12, 161 (2020). https://doi.org/10.1186/s13148-020-00940-3

Scrucca L., Fop M., Murphy T. B. and Raftery A. E. (2016) mclust 5:
clustering, classification and density estimation using Gaussian finite
mixture models The R Journal 8/1, pp. 289-317
}
