% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{clusterComparison}
\alias{clusterComparison}
\title{Compares the results of the two Clustering functions}
\usage{
clusterComparison(clusterAnalysisRes, clusterAnalysisTwoRes)
}
\arguments{
\item{clusterAnalysisRes}{A kmeans object, essentially a list with the
results from the clusterAnalysis function.}

\item{clusterAnalysisTwoRes}{A list with the results from the
clusterAnalysisTwo function.}
}
\value{
Returns a character vector of which contains the difference between
the two maximum clusters obtained from the clusterAnalysis and
clusterAnalysisTwo.
}
\description{
A function that compares the results of clusterAnalysis and
clusterAnalysisTwo by looking at the samples clustered in the maximum cluster
by both functions and getting the difference in the clusters.
}
\examples{
# Example 1:
# Using methylationData and significantSites Data available with package
dim(methylationData)
dim(significantSites)

# Do clustering Analysis
resultsOne <- clusterAnalysis(dnaMethylationData = methylationData, dnaMethylationSites = significantSites)
resultsTwo <- clusterAnalysisTwo(dnaMethylationData = methylationData, dnaMethylationSites = significantSites)

# Do cluster comparison
clusterComparison(clusterAnalysisRes = resultsOne, clusterAnalysisTwoRes = resultsTwo)

}
\references{
Wickham H, François R, Henry L, Müller K (2022). \emph{dplyr: A Grammar of Data
Manipulation}. R package version 1.0.10, \url{https://CRAN.R-project.org/package=dplyr}.

R Core Team (2022). R: A language and environment for statistical computing. R
Foundation for Statistical Computing, Vienna, Austria. URL
https://www.R-project.org/.
}
