% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusterPlot}
\alias{clusterPlot}
\title{Plot the k-means cluster results}
\usage{
clusterPlot(dnaMethylationData, dnaMethylationSites, clusterResults)
}
\arguments{
\item{dnaMethylationData}{A matrixwhich contains DNA Methylation beta
values of various methylation sites from several different samples.}

\item{dnaMethylationSites}{A matrix that contains the DNA methylation sites
which have been validated as being significant for the purposes of
clustering.}

\item{clusterResults}{A list which are the results from the k-means cluster.}
}
\value{
Returns a cluster plot.
}
\description{
A function that plots the results from the k-means cluster results.
}
\examples{
# Example 1:
# Using methylationData and significantSites Data available with package
dim(methylationData)
dim(significantSites)

# Do clustering Analysis
results <- clusterAnalysis(dnaMethylationData = methylationData, dnaMethylationSites = significantSites)

clusterPlot(dnaMethylationData = methylationData, dnaMethylationSites = significantSites, clusterResults = results)

}
\references{
Kassambara A, Mundt F (2020). \emph{factoextra: Extract and Visualize the
Results of Multivariate Data Analyses}. R package version 1.0.7,
\url{https://CRAN.R-project.org/package=factoextra}.
}
