% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compareClusterPlot}
\alias{compareClusterPlot}
\title{Compare Cluster Results}
\usage{
compareClusterPlot(clusterResults, clusterResultsTwo)
}
\arguments{
\item{clusterResults}{A list which are the results from clusterAnalysis}

\item{clusterResultsTwo}{A list which are the results from clusterAnalysisTwo}
}
\value{
Returns a figure with two barplots
}
\description{
A function that plots the barplots from the results of both clusterAnalysis
and clusterAnalysisTwo
}
\examples{
# Example 1:
# Using methylationData and significantSites Data available with package
dim(methylationData)
dim(significantSites)

# Do clustering Analysis
results <- clusterAnalysis(dnaMethylationData = methylationData, dnaMethylationSites = significantSites)
resultsTwo <- clusterAnalysisTwo(dnaMethylationData = methylationData, dnaMethylationSites = significantSites)

# Use the function to compare the clusters
compareClusterPlot(clusterResults = results, clusterResultsTwo = resultsTwo)

}
\references{
R Core Team (2022). R: A language and environment for statistical computing. R Foundation for Statistical
Computing, Vienna, Austria. URL https://www.R-project.org/.
}
