% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylationData}
\alias{methylationData}
\title{DNA methylation Beta values of various Gastric Cancer Patients.}
\format{
A matrix with columns:
\describe{
\item{ID}{The identity of the Methylation site of the DNA}
\item{TCGA-BR-4184-11A}{Identifier of sample used by TCGA.}
\item{TCGA-BR-4253-11A}{Identifier of sample used by TCGA.}
\item{TCGA-BR-4184-01A}{Identifier of sample used by TCGA.}
\item{TCGA-BR-4187-01A}{Identifier of sample used by TCGA.}
\item{TCGA-CG-4474-01A}{Identifier of sample used by TCGA.}
\item{TCGA-BR-4201-01A}{Identifier of sample used by TCGA.}
\item{TCGA-BR-4370-11A}{Identifier of sample used by TCGA.}
}
}
\source{
The Cancer Genome Atlas. TCGA
}
\usage{
methylationData
}
\description{
DNA Methylation Beta values collected from various methylation sites of
Gastric Cancer Patients.
}
\examples{
\dontrun{
 methylationData
}
}
\keyword{datasets}
