% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{significantSites}
\alias{significantSites}
\title{Significant DNA methylation sites of various Gastric Cancer Patients.}
\format{
A matrix with columns:
\describe{
\item{ID}{First condition, under which least light was provided
to plants.}
\item{Hazard ratio}{Hazard ratio calculated using survival analysis.}
\item{Low.95\% Confidence interval}{The low 95\% confidence Interval}
\item{High.95\% Confidence interval}{High 95\% confidence Interval}
\item{pvalue}{p-value}
}
}
\source{
Changhai Hospital, Navy Military Medical University, Shanghai 200433, China
}
\usage{
significantSites
}
\description{
Significant sites were determined using cox regression analysis.
}
\examples{
\dontrun{
 significantSites
}
}
\references{
Li, T., Chen, X., Gu, M., Deng, A., & Qian, C. (2020).
Identification of the subtypes of gastric cancer based on DNA methylation
and the prediction of prognosis. Clinical Epigenetics, 12(1), 161–161.
https://doi.org/10.1186/s13148-020-00940-3
}
\keyword{datasets}
